[None][test] Share the isinstance-transparent seam shim with the prefetcher
Move _isinstance_transparent_shim to _session_utils and use it at the
prefetcher's seam install too: when the prefetcher owns the seams
(reuse disabled or manual opt-in), it had the same hazard the shim
removes for reuse — a bare function at the seam makes any
isinstance(x, MpiPoolSession) against the patched module attribute
raise TypeError (the #16338 breakage class). Adds an install-side test
asserting the patched attribute stays a real type that answers for the
real class.
